# docs/swagger.yaml
basePath: /
definitions:
  api_teklifYonetimi_internal_api_response.APIResponse:
    properties:
      code:
        type: string
      data: {}
      message:
        type: string
      meta: {}
      success:
        type: boolean
    type: object
  api_teklifYonetimi_internal_dto.CreateCompanyRequest:
    properties:
      logo:
        type: string
      name:
        type: string
    required:
    - name
    type: object
...
  api_teklifYonetimi_internal_dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
...
  api_teklifYonetimi_internal_dto.UpdateCompanyRequest:
    properties:
      logo:
        type: string
      name:
        type: string
      subscription:
        type: string
    type: object
...
host: localhost:8082
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Teklif Yönetim Sistemi API Dokümantasyonu
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Teklif Yönetimi API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Email ve şifre ile giriş yapar, JWT token döner
      parameters:
      - description: Giriş Bilgileri
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/api_teklifYonetimi_internal_dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_teklifYonetimi_internal_api_response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_teklifYonetimi_internal_api_response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_teklifYonetimi_internal_api_response.APIResponse'
      summary: Kullanıcı Girişi
      tags:
      - Auth
  /companies:
    get:
      consumes:
      - application/json
      description: Tüm şirketleri listeler
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_teklifYonetimi_internal_api_response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api_teklifYonetimi_internal_api_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Şirketleri Listele
      tags:
      - Companies
    post:
      consumes:
      - application/json
      description: Yeni bir şirket kaydı oluşturur
      parameters:
      - description: Şirket Bilgileri
        in: body
        name: company
        required: true
        schema:
          $ref: '#/definitions/api_teklifYonetimi_internal_dto.CreateCompanyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api_teklifYonetimi_internal_api_response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_teklifYonetimi_internal_api_response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api_teklifYonetimi_internal_api_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Yeni Şirket Oluştur
      tags:
      - Companies
  /companies/{id}:
    delete:
      consumes:
      - application/json
      description: Şirketi siler
      parameters:
      - description: Şirket ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_teklifYonetimi_internal_api_response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api_teklifYonetimi_internal_api_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Şirket Sil
      tags:
      - Companies
    get:
      consumes:
      - application/json
      description: ID'ye göre şirket detayını getirir
      parameters:
      - description: Şirket ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_teklifYonetimi_internal_api_response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api_teklifYonetimi_internal_api_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Şirket Detayı
      tags:
      - Companies
    put:
      consumes:
      - application/json
      description: Şirket bilgilerini günceller
      parameters:
      - description: Şirket ID
        in: path
        name: id
        required: true
        type: integer
      - description: Güncelleme Bilgileri
        in: body
        name: company
        required: true
        schema:
          $ref: '#/definitions/api_teklifYonetimi_internal_dto.UpdateCompanyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_teklifYonetimi_internal_api_response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_teklifYonetimi_internal_api_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Şirket Güncelle
      tags:
      - Companies
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"